Audio and styles
Linked mp3 ogg, embeded an mp3 and said if the file is supported for both audio controls for Robert Frost Fire and Ice and for his poem Devotion. Then added styles to display the audio as block, have a box shadow, centered it and set the width to 80%, then changed the colors o f the control.